+++ fathom-integration-test/src/test/java/controllers/ApiControllerTest.java
@@ -16,13 +16,14 @@
 
 package controllers;
 
-import fathom.test.FathomIntegrationTest;
+import fathom.test.RestIntegrationTest;
+import models.Item;
 import org.junit.Test;
 
 import static com.jayway.restassured.RestAssured.given;
 import static org.hamcrest.Matchers.equalTo;
 
-public class ApiControllerTest extends FathomIntegrationTest {
+public class ApiControllerTest extends RestIntegrationTest {
 
     @Test
     public void testGetJSON() {
@@ -49,5 +50,20 @@ public class ApiControllerTest extends FathomIntegrationTest {
 
         given().accept(XML).when().get("/api/v1/items/{id}", 1).then().body("item.@id", equalTo("1"));
 
+    }
+
+    @Test
+    public void testGetObject() {
+
+        // The JSON response should look like:
+        //
+        // {
+        //   "id" : 1,
+        //   "name" : "Item 1"
+        // }
+
+        Item item = given().accept(JSON).when().get("/api/v1/items/{id}", 1).as(Item.class);
+        assertEquals("Item id does not match", 1, item.getId());
+        assertEquals("Item name does not match", "Apples", item.getName());
     }
 }